+++ app/assets/javascripts/govuk_publishing_components/dependencies.js
@@ -1,12 +1,9 @@
-/* eslint-env jquery */
+/* eslint-env */
 
 // This adds in javascript that initialises components and dependencies
 // that are provided by Slimmer in public frontend applications.
-// = require jquery/dist/jquery
 // = require ./modules.js
 
-$(document).ready(function () {
-  'use strict'
-
+document.addEventListener('DOMContentLoaded', function () {
   window.GOVUK.modules.start()
 })